refactor statistics into modular package
