+++ m3c2/core/statistics/__init__.py
@@ -1,0 +1,9 @@
+from .service import StatisticsService
+from .exporters import CANONICAL_COLUMNS, write_table, write_cloud_stats
+
+__all__ = [
+    "StatisticsService",
+    "CANONICAL_COLUMNS",
+    "write_table",
+    "write_cloud_stats",
+]
